board 에서 reply로 코드 옮기기
