+++ dokky/src/main/java/org/my/service/BoardServiceImpl.java
@@ -232,61 +232,4 @@ public class BoardServiceImpl implements BoardService {
 		return mapper.getDisLikeCount(num);
 	}
 	
-	@Override
-	public String checkReplyLikeValue(ReplyLikeVO vo) { 
-		
-		log.info("checkReplyLikeValue");
-		return mapper.checkReplyLikeValue(vo); 
-	}
-	
-	@Transactional
-	@Override
-	public int registerReplyLike(ReplyLikeVO vo) {//댓글 좋아요 컬럼 등록 및 좋아요 push
-
-		log.info("registerReplyLike...." + vo);
-		
-		mapper.registerReplyLike(vo);
-		
-		log.info("pushReplyLike...."+vo.getReply_num());
-		
-		return mapper.pushReplyLike(vo.getReply_num()); 
-	}
-	
-	@Transactional
-	@Override
-	public int pushReplyLike(ReplyLikeVO vo) {//댓글 좋아요 누르기  
-		
-		log.info("pushLikeValue...."+vo);  
-		
-		mapper.pushReplyLikeValue(vo);
-		
-		log.info("pushReplyLike...."+vo.getReply_num());
-		
-		return mapper.pushReplyLike(vo.getReply_num()); 
-	}
-	
-	@Transactional 
-	@Override
-	public int pullReplyLike(ReplyLikeVO vo) {//댓글  좋아요 취소 pull
-		
-		log.info("pullReplyLikeValue...."+vo);
-		
-		mapper.pullReplyLikeValue(vo);
-		
-		log.info("pullReplyLike...."+vo.getReply_num());
-		
-		return mapper.pullReplyLike(vo.getReply_num());
-	}
-	
-	@Override
-	public String getReplyLikeCount(Long reply_num) {
-  
-		log.info("getReplyLikeCount");
-		return mapper.getReplyLikeCount(reply_num);
-	}
-	
-	
-	
-	
-	
 }
