+++ dokky/src/main/java/org/my/service/ReplyService.java
@@ -1,7 +1,8 @@
 package org.my.service;
 
 	import org.my.domain.Criteria;
-	import org.my.domain.ReplyPageDTO;
+import org.my.domain.ReplyLikeVO;
+import org.my.domain.ReplyPageDTO;
 	import org.my.domain.ReplyVO;
 
 public interface ReplyService {
@@ -15,5 +16,15 @@ public interface ReplyService {
 	public int remove(Long reply_num);
 	
 	public ReplyPageDTO getListPage(Criteria cri, Long num);
+	
+	public String checkReplyLikeValue(ReplyLikeVO vo);
+
+	public int registerReplyLike(ReplyLikeVO vo);
+
+	public int pushReplyLike(ReplyLikeVO vo);
+
+	public int pullReplyLike(ReplyLikeVO vo);
+
+	public String getReplyLikeCount(Long reply_num);
 
 }
