+++ dokky/src/main/webapp/WEB-INF/views/board/get.jsp
@@ -494,7 +494,7 @@
 		replyService.updateReplyLike(likeData, function(result){
 		 
 		var replyLikeCount = $("#replyLikeCount"+reply_num);
-		replyLikeCount.html(result);
+			replyLikeCount.html(result);
 			//console.log(result); 
 		  });
 		//추후 좋아요를 눌르면 이미지변경까지 취소하면 이미지변경 추가해보자
